+++ src/components/Persons/Persons.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Person from './Person/Person';
+
+const persons = (props) => props.persons.map((person, index) => {
+    return <Person
+        name={person.name}
+        age={person.age}
+        key={person.id}
+        changed={(event) => props.changed(event, person.id)}
+        clicked={() => props.clicked(index)}
+    />
+})
+
+export default persons;
